🚚 move custom incremental cache handler path to use ESM
