+++ package.json
@@ -42,7 +42,7 @@
     "geist": "^1.0.0",
     "isbot": "^3.6.13",
     "nanoid": "^4.0.2",
-    "next": "14.0.4",
+    "next": "14.0.5-canary.38",
     "nextjs-toploader": "^1.6.4",
     "nprogress": "^0.2.0",
     "pg": "^8.11.3",
